Additional Crash Fixes for ExitCrafting
Merge in from the previous build I'd been working on before 7.1 (as I had this crash again in 7.1 from the previous compile). Two new functions and alternate handling in ExitCrafting to use the macro if currently crafting, the RecipeNote.Close function if the UI is just open but not crafting, and not try anything if the UI isn't open. This seems to be more stable than the old behavior.